+++ resources/webviews/azureserviceoperator/azureserviceoperator.css
@@ -32,7 +32,7 @@ body {
     border-radius: 2px;
 }
 
-.container .allocatebody, .asooutput {
+.container .allocatebody, .asooutput .asoform {
     border-color: #ddd;
     border-left-width: 4px;
     padding-left: 20px;
@@ -71,4 +71,20 @@ th {
 
 label {
     cursor: pointer;
+}
+
+/* Loader icon */
+.loader {
+  border: 16px solid #f3f3f3; /* Light grey */
+  border-top: 16px solid #3498db; /* Blue */
+  border-radius: 50%;
+  width: 120px;
+  height: 120px;
+  align-content: center;
+  animation: spin 2s linear infinite;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
 }
