+++ resources/webviews/azureserviceoperator/azureserviceoperator.html
@@ -8,6 +8,7 @@
     <!-- Link to the css file -->
     <link rel="stylesheet" href="{{cssuri}}">
     <script src="https://use.fontawesome.com/7dddd54b5c.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.2/jquery.validate.min.js"></script>
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"
         integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
 
@@ -17,14 +18,24 @@
 
         $(document).ready(function () {
             $('.asooutput').hide();
-
-            $(".toggle").click(function () {
-                $('.allocatebody').toggle('slow');
-            });
+            $('.loader').hide();
 
             $(".toggleoutput").click(function () {
                 $('.asooutput').toggle('slow');
             });
+
+            $('#spsubmit').click(function () {
+                var appid = $('#spappid').val();
+                var password = $('#spcred').val();
+
+                if (appid.length > 0 && password.length > 0) {
+                    $('.loader').show();
+                    $('.spform').hide();
+                    vscode.postMessage({ appid: appid, password: password });
+                }
+            })
+
+
         });
 
     </script>
@@ -44,7 +55,35 @@
             troubleshooting process.
             <a href="https://operatorhub.io/operator/azure-service-operator">Learn more</a></a>
     </div>
-
+    {{#if getUserInput}}
+    <div class="container">
+        <div class="loader"></div>
+        <div class="panel panel-default data-container spform">
+            <i class="fa fa-wpforms" aria-hidden="true"
+                style="font-size:14px;padding-right:5px;padding-bottom:15px;"></i>
+            Please provide service principal appid and password, as detailed in <a
+                href="https://operatorhub.io/operator/azure-service-operator">step 4 of azure service operator docs</a>.
+            <br />
+            <blockquote cite="https://operatorhub.io/operator/azure-service-operator">
+                Service Principal with Contributor permissions for your subscription is required,
+                so ASO can create resources in your subscription on your behalf.
+                Note that the <a
+                    href="https://docs.microsoft.com/en-us/cli/azure/create-an-azure-service-principal-azure-cli">ServicePrincipal</a>
+                you pass to the command below needs to have
+                access to create resources in your subscription.
+            </blockquote>
+            <form id="serviceprincipalform" class="allocatebody">
+                <label for="spappid">Enter App ID of service principal:</label>
+                <input required id="spappid" name="spappid" size="50" type="text"
+                    placeholder="e.g. 041ccd53-e72f-45d1-bbff-382c82f6f9a1" /><br><br>
+                <label for="spcred">Enter Password of Service Principal:</label>
+                <input required id="spcred" name="spcred" size="50" type="text"
+                    placeholder="Service principal password" /><br><br>
+                <input id="spsubmit" type="submit" value="Submit">
+            </form>
+        </div>
+    </div>
+    {{else}}
     <div class="container">
         {{#if isSuccess}}
         <div class="panel panel-default data-container">
@@ -76,7 +115,7 @@
             </div>
         </div>
     </div>
-
+    {{/if}}
 </body>
 
 </html>
