+++ src/commands/azureServiceOperators/helpers/azureservicehelper.ts
@@ -1,51 +1,11 @@
 import * as vscode from 'vscode';
 import * as k8s from 'vscode-kubernetes-tools-api';
-import * as htmlhandlers from "handlebars";
 import * as path from 'path';
 import * as fs from 'fs';
 import { getExtensionPath } from '../../utils/host';
 import { OperatorSettings } from '../models/operatorSettings';
-import AksClusterTreeItem from '../../../tree/aksClusterTreeItem';
 import * as tmpfile from '../../utils/tempfile';
-import { InstallationResponse } from '../models/installationResponse';
 const tmp = require('tmp');
-
-export async function getAzureServicePrincipal(
-    aksCluster: AksClusterTreeItem
-): Promise<OperatorSettings | undefined> {
-
-    const optionsAppId: vscode.InputBoxOptions = {
-        prompt: "Enter AppId of Service Principal, the ServicePrincipal you pass to the command below needs to have access to create resources in your subscription. More: https://docs.microsoft.com/en-us/cli/azure/create-an-azure-service-principal-azure-cli",
-        placeHolder: "Enter App ID of service principal: e.g.(Sample) 041ccd53-e72f-45d1-bbff-382c82f6f9a1",
-        ignoreFocusOut: true
-    };
-    const optionsPassword: vscode.InputBoxOptions = {
-        prompt: "Enter Password for the Service Principal, the ServicePrincipal you pass to the command below needs to have access to create resources in your subscription. More: https://docs.microsoft.com/en-us/cli/azure/create-an-azure-service-principal-azure-cli",
-        placeHolder: "Enter Password of Service Principal",
-        ignoreFocusOut: true
-    };
-
-    const inputAppIdBox = await vscode.window.showInputBox(optionsAppId);
-    const inputPasswordBox = await vscode.window.showInputBox(optionsPassword);
-    const cloudName = convertAzureCloudEnv(aksCluster.root.environment.name);
-
-    if (!inputAppIdBox || !inputPasswordBox) {
-        return undefined;
-    }
-
-    if (!cloudName) {
-        vscode.window.showWarningMessage(`Cloud environment name ${cloudName} is not supported.`);
-        return undefined;
-    }
-
-    return {
-        tenantId: aksCluster.root.tenantId,
-        subId: aksCluster.subscription.subscriptionId!,
-        appId: inputAppIdBox,
-        clientSecret: inputPasswordBox,
-        cloudEnv: cloudName
-    };
-}
 
 export async function getOperatorsPod(
     kubectl: k8s.KubectlV1,
@@ -180,163 +140,4 @@ async function invokeKubectlCommand(
         vscode.window.showErrorMessage(`${failureDescription}: ${e}`);
         return undefined;
     }
-}
-
-function convertAzureCloudEnv(cloudName: string): string | undefined {
-    // Cloud env map: https://docs.microsoft.com/en-us/azure/storage/common/storage-powershell-independent-clouds#get-endpoint-using-get-azenvironment
-    if (cloudName === "AzureUSGovernment") {
-        return "AzureUSGovernmentCloud";
-    }
-    if (cloudName === "AzureChinaCloud") {
-        return "AzureChinaCloud";
-    }
-    if (cloudName === "AzureGermanCloud") {
-        return "AzureGermanCloud";
-    }
-    if (cloudName === "AzureCloud") {
-        return "AzurePublicCloud";
-    }
-
-    return undefined;
-}
-
-interface InstallResult {
-    succeeded?: boolean;  // used for how to format the main message
-    mainMessage?: string;  // the "it installed" or "it failed" message
-    logs?: LogSection[];  // the logs
-}
-
-interface LogSection {
-    title?: string;
-    messages?: string;
-}
-
-export function getWebviewContent(
-    clustername: string,
-    aksExtensionPath: string,
-    installationResponse: InstallationResponse
-): string {
-    const stylePathOnDisk = vscode.Uri.file(path.join(aksExtensionPath, 'resources', 'webviews', 'azureserviceoperator', 'azureserviceoperator.css'));
-    const htmlPathOnDisk = vscode.Uri.file(path.join(aksExtensionPath, 'resources', 'webviews', 'azureserviceoperator', 'azureserviceoperator.html'));
-    const styleUri = stylePathOnDisk.with({ scheme: 'vscode-resource' });
-    const pathUri = htmlPathOnDisk.with({ scheme: 'vscode-resource' });
-
-    const htmldata = fs.readFileSync(pathUri.fsPath, 'utf8').toString();
-    const installHtmlResult = getOrganisedInstallResult(clustername, installationResponse);
-
-    htmlHandlerRegisterHelper();
-    const template = htmlhandlers.compile(htmldata);
-    const data = {
-        cssuri: styleUri,
-        name: clustername,
-        mainMessage: installHtmlResult.mainMessage,
-        resultLogs: installHtmlResult.logs,
-        isSuccess: installHtmlResult.succeeded
-    };
-    const webviewcontent = template(data);
-
-    return webviewcontent;
-}
-
-function getOrganisedInstallResult(
-    clustername: string,
-    installationResponse: InstallationResponse
-) {
-    const installResults: InstallResult = {};
-    const certManagerResult = installationResponse.installCertManagerResult;
-    const certManagerSatatusResult = installationResponse.checkCertManagerRolloutStatusResult;
-    const issuerCertResult = installationResponse.installIssuerCertResult;
-    const olmCrdResult = installationResponse.installOlmCrdResult;
-    const olmResult = installationResponse.installOlmResult;
-    const operatorResult = installationResponse.installOperatorSettingsResult;
-    const operatorSettingsResult = installationResponse.installOperatorResult;
-    const getOperatorsPodResult = installationResponse.getOperatorsPodResult;
-
-    const installResultCollection = [certManagerResult, certManagerSatatusResult, issuerCertResult, olmCrdResult,
-        olmResult, operatorResult, operatorSettingsResult, getOperatorsPodResult];
-    installResults.succeeded = installationExitCodeResponse(installResultCollection);
-
-    if (installResults.succeeded) {
-        installResults.mainMessage = `Azure Service Operator Successfully installed on ${clustername}. Please see the console output below for more details.`;
-    } else {
-        installResults.mainMessage = `Azure Service Operator Failed to install on ${clustername}. Please see the console output below for more details.`;
-    }
-
-    const logs = [];
-
-    if (certManagerResult) {
-        logs.push(getHtmlLogSectionFromResponse("Cert Manager Output", certManagerResult));
-    }
-
-    if (certManagerSatatusResult) {
-        logs.push(getHtmlLogSectionFromResponse("Issuer Cert RollOut Status", certManagerSatatusResult));
-    }
-
-    if (issuerCertResult) {
-        logs.push(getHtmlLogSectionFromResponse("Issuer Cert Output", issuerCertResult));
-    }
-
-    if (olmCrdResult) {
-        logs.push(getHtmlLogSectionFromResponse("Install Operator Lifecycle CRD Output", olmCrdResult));
-    }
-
-    if (olmResult) {
-        logs.push(getHtmlLogSectionFromResponse("Install Operator Lifecycle Output", olmResult));
-    }
-
-    if (operatorResult) {
-        logs.push(getHtmlLogSectionFromResponse("Install Operator Output", operatorResult));
-    }
-
-    if (operatorSettingsResult) {
-        logs.push(getHtmlLogSectionFromResponse("ASO Setting Apply Output", operatorSettingsResult));
-    }
-
-    if (getOperatorsPodResult) {
-        logs.push(getHtmlLogSectionFromResponse("Operator pod output", getOperatorsPodResult));
-    }
-
-    installResults.logs = logs;
-
-    return installResults;
-}
-
-function getHtmlLogSectionFromResponse(
-    title: string,
-    shellResult: k8s.KubectlV1.ShellResult
-): LogSection {
-    const logSection: LogSection = {};
-
-    if (shellResult) {
-        logSection.title = title;
-        logSection.messages = installationStdResponse(shellResult);
-    }
-
-    return logSection;
-}
-
-function installationStdResponse(
-    installShellResult: k8s.KubectlV1.ShellResult | undefined
-): string | undefined {
-    // Standard output code of all the installation response.
-    return installShellResult ? installShellResult.stderr + installShellResult.stdout : undefined;
-}
-
-function installationExitCodeResponse(
-    installShellResult: (k8s.KubectlV1.ShellResult | undefined)[],
-): boolean {
-    const isSuccess = installShellResult.filter(Boolean).every((sr) => !!sr && sr.code === 0);
-
-    return isSuccess;
-}
-
-export function htmlHandlerRegisterHelper() {
-    htmlhandlers.registerHelper("breaklines", breaklines);
-}
-
-function breaklines(text: any): any {
-    if (text) {
-        text = text.replace(/(\r\n|\n|\r)/gm, '<br>');
-    }
-    return text;
 }
